+++ client/web/src/components/MapPage/index.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { useLocation } from 'react-router-dom'
 import * as select from 'store/selectors'
-import { saveQuery, getJurisdiction, clearJurisdiction } from 'store/actions'
+import { getJurisdiction, clearJurisdiction } from 'store/actions/data'
+import { saveQuery } from 'store/actions/query'
 import useBreakpoints from 'hooks/useBreakpoints'
 import Desktop from './Desktop'
 import Mobile from './Mobile'
