+++ client/web/src/components/MapPage/shared/JurisdictionSelect/index.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { useHistory } from 'react-router-dom'
 import queryString from 'query-string'
 import * as select from 'store/selectors'
-import { getStatesWithJurisdictions } from 'store/actions'
+import { getStatesWithJurisdictions } from 'store/actions/data'
 import { makeStyles } from '@material-ui/core/styles'
 import useBreakpoints from 'hooks/useBreakpoints'
 import IconButton from '@material-ui/core/IconButton'
