+++ client/web/src/components/MapPage/shared/LocationDetail/CheckSteps/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { withStyles } from '@material-ui/core/styles'
 import IconButton from '@material-ui/core/IconButton'
-import { openModal } from 'store/actions'
+import { openModal } from 'store/actions/modals'
 import Accordion from './Accordion'
 import { ReactComponent as CheckCircleIcon } from 'assets/icons/check-circle.svg'
 import { ReactComponent as ArrowForwardIcon } from 'assets/icons/arrow-forward.svg'
