+++ client/web/src/components/MapPage/shared/LocationList/LocationCards.js
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import * as select from 'store/selectors'
-import { selectLocation } from 'store/actions'
+import { selectLocation } from 'store/actions/query'
 import LocationCard from '../LocationCard'
 import Divider from '@material-ui/core/Divider'
 import { makeStyles } from '@material-ui/core/styles'
