+++ client/web/src/components/MapPage/shared/Map/index.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useCallback } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import * as select from 'store/selectors'
-import { selectLocation } from 'store/actions'
+import { selectLocation } from 'store/actions/query'
 import { lineString } from '@turf/helpers'
 import bbox from '@turf/bbox'
 import Map from './Map'
