+++ client/web/src/components/main/Header.jsx
@@ -10,7 +10,7 @@ import logo from 'assets/logos/ballotnav.svg'
 import Footer from 'components/main/Footer'
 import IconButton from '@material-ui/core/IconButton'
 import SearchIcon from '@material-ui/icons/Search'
-import { openModal } from 'store/actions'
+import { openModal } from 'store/actions/modals'
 import * as select from 'store/selectors'
 import { makeStyles } from '@material-ui/core/styles'
 import Div100vh from 'react-div-100vh'
