+++ client/web/src/components/modals/SearchModal.js
@@ -6,7 +6,7 @@ import Dialog from '@material-ui/core/Dialog'
 import IconButton from '@material-ui/core/IconButton'
 import ArrowBackIcon from '@material-ui/icons/ArrowBack'
 import * as select from 'store/selectors'
-import { closeModal } from 'store/actions'
+import { closeModal } from 'store/actions/modals'
 import SearchBar from 'components/shared/SearchBar'
 import JurisdictionSelect from 'components/MapPage/shared/JurisdictionSelect'
