+++ client/web/src/components/modals/ShareModal.js
@@ -2,7 +2,7 @@ import React, { useMemo, useCallback, useState } from 'react'
 import { connect } from 'react-redux'
 import { useLocation } from 'react-router-dom'
 import * as select from 'store/selectors'
-import { closeModal } from 'store/actions'
+import { closeModal } from 'store/actions/modals'
 import Dialog from '@material-ui/core/Dialog'
 import IconButton from '@material-ui/core/IconButton'
 import CloseIcon from '@material-ui/icons/Close'
