+++ client/web/src/components/modals/VoteDotOrgModal.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import * as select from 'store/selectors'
-import { closeModal } from 'store/actions'
+import { closeModal } from 'store/actions/modals'
 import Dialog from '@material-ui/core/Dialog'
 import IconButton from '@material-ui/core/IconButton'
 import CloseIcon from '@material-ui/icons/Close'
