+++ client/web/src/store/index.js
@@ -1,8 +1,16 @@
-import { createStore, applyMiddleware } from 'redux'
+import { combineReducers, createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly'
 
-import rootReducer from './actions'
+import query from './actions/query'
+import data from './actions/data'
+import modals from './actions/modals'
+
+const rootReducer = combineReducers({
+  query,
+  data,
+  modals,
+})
 
 const store = createStore(
   rootReducer,
